Move contracts.go and contracts_test.go to project
