--- flow/beta/beta.go
@@ -1,36 +1,0 @@
-/*
- * Flow CLI
- *
- * Copyright 2019-2020 Dapper Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package beta
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/onflow/flow-cli/flow/beta/commands/deploy_contracts"
-	"github.com/onflow/flow-cli/flow/beta/commands/start_emulator"
-)
-
-var Cmd = &cobra.Command{
-	Use:              "beta",
-	TraverseChildren: true,
-}
-
-func init() {
-	Cmd.AddCommand(start_emulator.Cmd)
-	Cmd.AddCommand(deploy_contracts.Cmd)
-}
