--- flow/beta/cli/cli.go
@@ -1,56 +1,0 @@
-/*
- * Flow CLI
- *
- * Copyright 2019-2020 Dapper Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cli
-
-import (
-	"crypto/rand"
-	"fmt"
-	"os"
-
-	"github.com/fatih/color"
-)
-
-const EnvPrefix = "FLOW"
-
-func Exit(code int, msg string) {
-	fmt.Println(msg)
-	os.Exit(code)
-}
-
-func Exitf(code int, msg string, args ...interface{}) {
-	fmt.Printf(msg+"\n", args...)
-	os.Exit(code)
-}
-
-func RandomSeed(n int) []byte {
-	seed := make([]byte, n)
-
-	_, err := rand.Read(seed)
-	if err != nil {
-		Exitf(1, "Failed to generate random seed: %v", err)
-	}
-
-	return seed
-}
-
-// Colors
-var Yellow = color.New(color.FgYellow, color.Bold).SprintfFunc()
-var Green = color.New(color.FgGreen, color.Bold).SprintfFunc()
-var Red = color.New(color.FgRed, color.Bold).SprintfFunc()
-var Bold = color.New(color.Bold).SprintfFunc()
